// src/postion/utils/postion.enum.ts
export enum Postion{
   
    ADMIN = 'ADMIN',
    SUPERADMIN = 'SUPERADMIN',
    BRANCHMANAGER = 'BRANCHMANAGER',
    COORDINATOR = 'COORDINATOR',
    RECEPTIONIST = 'RECEPTIONIST',
    ACCOUNTANT = 'ACCOUNTANT',
    ARTIST = 'ARTIST',
    ARTISTMANAGER = 'ARTISTMANAGER',
    TABLEMANAGER = 'TABLEMANAGER'


}